Rename my component folders again, to try and make macOS recognize the fact that I renamed them.
